refactor(watchlist): define dedicated action component
