+++ projects/client/src/lib/sections/actions/watchlist/useWatchlist.ts
@@ -3,9 +3,9 @@ import type { MediaType } from '$lib/models/MediaType.ts';
 import { InvalidateAction } from '$lib/requests/models/InvalidateAction.ts';
 import { addToWatchlistRequest } from '$lib/requests/sync/addToWatchlistRequest.ts';
 import { removeFromWatchlistRequest } from '$lib/requests/sync/removeFromWatchlistRequest.ts';
+import { toWatchlistPayload } from '$lib/sections/actions/watchlist/toWatchlistPayload';
+import { useInvalidator } from '$lib/stores/useInvalidator.ts';
 import { derived, writable } from 'svelte/store';
-import { toWatchlistPayload } from './_internal/toWatchlistPayload.ts';
-import { useInvalidator } from './useInvalidator.ts';
 
 export type WatchlistStoreProps = {
   type: MediaType;
